Added wreath recipe, moved events to the correct bus, reformatted code , fixed classloading errors on the server and removed useless eventbussubscriber annotations.
