+++ src/main/java/com/lilypuree/decorative_winter/blocks/ModBlockProperties.java
@@ -3,5 +3,6 @@ package com.lilypuree.decorative_winter.blocks;
 import net.minecraft.state.IntegerProperty;
 
 public class ModBlockProperties {
+
     public static final IntegerProperty LAYERS_0_8 = IntegerProperty.create("layers", 0, 8);
 }
