+++ src/main/java/com/lilypuree/decorative_winter/blocks/SnowyPalisadeBlock.java
@@ -2,21 +2,9 @@ package com.lilypuree.decorative_winter.blocks;
 
 import com.lilypuree.decorative_blocks.blocks.PalisadeBlock;
 import com.lilypuree.decorative_blocks.datagen.types.IWoodType;
-import com.lilypuree.decorative_winter.setup.ModSetup;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
-import net.minecraft.block.Blocks;
-import net.minecraft.block.FenceGateBlock;
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.entity.item.ItemEntity;
-import net.minecraft.fluid.FluidState;
-import net.minecraft.fluid.Fluids;
-import net.minecraft.item.BlockItemUseContext;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.Items;
 import net.minecraft.state.BooleanProperty;
-import net.minecraft.state.IntegerProperty;
-import net.minecraft.state.Property;
 import net.minecraft.state.StateContainer;
 import net.minecraft.state.properties.BlockStateProperties;
 import net.minecraft.util.Direction;
@@ -26,15 +14,11 @@ import net.minecraft.util.math.shapes.VoxelShape;
 import net.minecraft.util.math.shapes.VoxelShapes;
 import net.minecraft.world.IBlockReader;
 import net.minecraft.world.IWorld;
-import net.minecraft.world.LightType;
-import net.minecraft.world.World;
-import net.minecraft.world.server.ServerWorld;
 
-import javax.annotation.Nullable;
 import java.util.Arrays;
-import java.util.Random;
 
 public class SnowyPalisadeBlock extends PalisadeBlock implements ISnowLoggable {
+
     public static final BooleanProperty SNOWY = BlockStateProperties.SNOWY;
     protected final VoxelShape[] snowyShapes;
 
@@ -46,9 +30,9 @@ public class SnowyPalisadeBlock extends PalisadeBlock implements ISnowLoggable {
 
     @Override
     public VoxelShape getShape(BlockState state, IBlockReader worldIn, BlockPos pos, ISelectionContext context) {
-        if (state.get(SNOWY)){
+        if (state.get(SNOWY)) {
             return snowyShapes[this.getIndex(state)];
-        }else {
+        } else {
             return super.getShape(state, worldIn, pos, context);
         }
     }
