+++ src/main/java/com/lilypuree/decorative_winter/blocks/SnowySeatBlock.java
@@ -2,7 +2,8 @@ package com.lilypuree.decorative_winter.blocks;
 
 import com.lilypuree.decorative_blocks.blocks.SeatBlock;
 import com.lilypuree.decorative_blocks.datagen.types.IWoodType;
-import net.minecraft.block.*;
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
 import net.minecraft.state.BooleanProperty;
 import net.minecraft.state.StateContainer;
 import net.minecraft.state.properties.BlockStateProperties;
@@ -15,6 +16,7 @@ import net.minecraft.world.IBlockReader;
 import net.minecraft.world.IWorld;
 
 public class SnowySeatBlock extends SeatBlock implements ISnowLoggable {
+
     private static final VoxelShape SNOW = Block.makeCuboidShape(0.0D, 0.0D, 0.0D, 16.0D, 2.0D, 16.0D);
     private static final VoxelShape POST_SHAPE_a = Block.makeCuboidShape(6.0D, 0.0D, 6.0D, 10.0D, 4.0D, 10.0D);
     private static final VoxelShape JOIST_SHAPE_NS_a = Block.makeCuboidShape(0.0D, 4.0D, 4.0D, 16.0D, 7.0D, 12.0D);
