+++ src/main/java/com/lilypuree/decorative_winter/client/model/RandomTranslatedModel.java
@@ -10,18 +10,17 @@ import net.minecraft.client.renderer.vertex.DefaultVertexFormats;
 import net.minecraft.util.Direction;
 import net.minecraft.util.math.MathHelper;
 import net.minecraft.util.math.vector.Vector2f;
-import net.minecraft.util.math.vector.Vector3d;
 import net.minecraftforge.client.model.data.IDynamicBakedModel;
 import net.minecraftforge.client.model.data.IModelData;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
 import java.util.stream.Collectors;
 
 public class RandomTranslatedModel implements IDynamicBakedModel {
+
     private IBakedModel baseModel;
 
     public RandomTranslatedModel(IBakedModel baseModel) {
